refactor: restore initial layout
